Avoid excessive list-runners call from runner controller in registration timeout check
Setting a short sync-period on actions-runner-controller had resulted in a series of paginated list-runners GitHub API calls for each sync on every idle runner. This was not necessarily. The original reason we wanted to call the GitHub API here was to determine if the runner needs to be recreated due to the registration timeout. A registration timeout happens only after 10 minutes after the pod creation so we do not need to check earlier than that. Since this change, the controller checks for the registration timeout only after approx 10 minutes regardless of how short the sync period is.