Optional override of runner image in chart
This commit adds the option to override the actions runner image. This
allows running the controller in environments where access to Dockerhub
is restricted.

It uses the parameter [--runner-image](https://github.com/actions-runner-controller/actions-runner-controller/blob/master/main.go#L89) from the controller.
The default value is set as a constant
[here](https://github.com/actions-runner-controller/actions-runner-controller/blob/acb906164b6cc8b3604fa816cd3a99acffe991a2/main.go#L40).

The default value for the chart is the same.