Add support for default image pull secret name
Currently it's supported to define imagePullSecrets in the runner spec
e.g as a RunnerDeployment. However the controller can specify a default
runner image.

If this image is on a private registry it requires the
image pull secret. This PR adds this option also to the controller to be
consistent with supporting a default runner image. This makes it much
easier when using runner images from private registries