Use Docker socket via shared emptyDir instead of TCP/mTLS
By default, the `docker:dind` entrypoint will auto-generate mTLS certs
and run with TCP on `0.0.0.0`. This is handy for accessing the running
Docker Engine remotely by then publishing the ports. For the runner,
we don't need (or want) that behavior, so a Unix socket lets us rely
on filesystem permissions.

This also has the benefit of eliminating the need for mTLS, which will
speed up Pod start slightly (no need to generate CA & client certs),
and will fix #893 and generally improve compatibility with apps that
interact with the Docker API without requiring a custom Docker context
to be initialized.