Fix docker.sock permission error for non-dind runners since v0.27.2
#2490 has been happening since v0.27.2 for non-dind runners based on Ubuntu 20.04 runner images. It does not affect Ubuntu 22.04 runners and Unbut 20.04/22.04 dind runners. Presuming many folks are still using Ubuntu 20.04 runners and non-dind runners, we should fix it.

This change tries to fix it by defaulting to the docker group id 1001 used by Ubuntu 20.04 runners, and use gid 121 for Ubuntu 22.04 runners. We use image tags to see which Ubuntu version the runner is basaed on.

This might be a breaking change for folks who had already upgraded to Ubuntu 22.04 runners using their own custom runner images. Note again, we rely on the image tag to detect Ubuntu 22.04 runner images and use the proper docker gid- Folks using our official Ubuntu 22.04 runner images are not affected. It is a breaking change anyway so I have added a remedy-

ARC got a new flag, `--docker-gid`, which defaults to `1001` but can be set to `121` or whatever gid the operator/admin likes. This can be set to `121`, for example, if you are using your own custom runner image based Ubuntu 22.04 and the image tag does not contain "22.04".

Fixes #2490