feat: allow for modifying `var-run` mount maximum size limit
When running a large number of containers in a single workflow job (via `docker-compose` for example) the current 1M size of the `var-run` volume is too small. This commit adds a new `dockerVarRunVolumeSizeLimit` parameter that allows users to customize the limit of memory for the `var-run` volume allowing to run more containers than with the defaults.

fixes actions/actions-runner-controller#2621