Fix `panic: slice bounds out of range` when runner spec contains `volumeMounts`.
As observed in #2719, if a `Runner` contains a `work` volume mount, this
will cause a panic in `newPod()` where the code is assuming that the
`work` volume mount will be present in the `pod` definition generated by
`newRunnerPodWithContainerMode()`.

However that isn't always the case, and if it's not present the `-1`
returned by `workVolumeMountPresent` results in the `slice bounds out of
range` exception.

So instead, check the `bool` return value, and only attempt to remove
the `work` volume mount if it's actually present.

Also apply the same fix to the `volumes` logic.

Fixes #2719

Signed-off-by: Gavin Williams <gavin.williams@machinemax.com>