Update values.yaml
When using this template, my pods were stuck in a restart loop. Providing the command field fixed the issue. Did not test the kubernetes option yet but I assume it needs the same start command.

Note that the docs here https://docs.github.com/en/actions/hosting-your-own-runners/managing-self-hosted-runners-with-actions-runner-controller/deploying-runner-scale-sets-with-actions-runner-controller#using-docker-in-docker-mode are also missing the command field