Facilitate detection of failed pod creation
Currently, the only way to see if the pod creation fails is to look at the runner_controller's log.
This requires permission to view the namespace where the runner controller is located.
Even if you do not have that permission, you should be able to see a log of a Runner you created.

Signed-off-by: utam0k <k0ma@utam0k.jp>