Use timestamps from Workflow Job event as defaults for queued,started,completed time
Default these values to the timestamps contained in the event. They will
be updated if the log contains more accurate values, however the
"Waiting for a runner to pick up this job..." and "Job is about to start
running on the runner:" lines are only present in the logs until the job
has finished, at which point they are removed from the job logs by
Github, so they can't be used to calculate the job duration after the
job has finished.

Fixes actions/actions-runner-controller#3522