Delete config secret when listener pod gets deleted
This should fix https://github.com/actions/actions-runner-controller/issues/4029.
When the github token gets updated, and the old one expires, the listener pod dies. Only to be recreated with the existing listener config that holds the now expired token.
This PR fixes that by deleting the config secret as well, so the reconciler recreates it with the updated token.