Use dumb-init from Ubuntu repositories
dumb-init is available in the Ubuntu package repositories, so let's use
that. This makes adding new architectures easier.

For Ubuntu 20.04, this means a downgrade from 1.2.5 to 1.2.2. However,
given that Ubuntu 20.04 reached the end of its support cycle in May
2025, this seems acceptable. The only regression is that dumb-init
won't attempt to cd into the root directory in order to release the file
descriptor for its current working directory, besides some fixes around
the --help and --version flags.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>